test: 포트폴리오 첨부 용량 검증 코드 위치 이동에 따른 테스트 코드 수정
